feat: add multipart upload handling and notification system
- Implemented new API routes for handling multipart uploads and notifications in the onboarding process.
- Added `POST` endpoint for creating part URLs for pending assets in both client and admin contexts.
- Introduced notification system to inform clients about uploaded files, including email formatting and sending logic.
- Enhanced rate limiting for upload and notification requests to prevent abuse.
- Created a new database migration to support multipart uploads by adding a `multipart_upload_id` column to the `onboarding_assets` table.

+++ lib/onboarding/ai-boundary.ts
@@ -30,7 +30,7 @@ export async function getWebsiteGenerationInput(projectId: string) {
       id,
       original_filename as name,
       mime_type as "mimeType",
-      size::int as size,
+      size::float8 as size,
       storage_key as "objectKey"
     from onboarding_assets
     where project_id = ${projectId} and status = 'uploaded'
